+++ SearchPanel.lua
@@ -732,13 +732,15 @@ local function updatePersistentResultFrame(resultID, isInviteFrame)
 			for i = 1, searchResultInfo.numMembers, 1 do
 				local role, class, _, specLocalized, isLeader = C_LFGList.GetSearchResultMemberInfo(searchResultInfo.searchResultID, i)
 
-				orderedList[i] = {leader = isLeader, role = role, class = class, specID = miog.LOCALIZED_SPECIALIZATION_NAME_TO_ID[specLocalized .. "-" .. class]}
+				if(class) then
+					orderedList[i] = {leader = isLeader, role = role, class = class, specID = miog.LOCALIZED_SPECIALIZATION_NAME_TO_ID[specLocalized .. "-" .. class]}
 
-				if(role) then
-					roleCount[role] = roleCount[role] + 1
+					if(role) then
+						roleCount[role] = roleCount[role] + 1
 
-				else
-					orderedList[i].role = "DAMAGER"
+					else
+						orderedList[i].role = "DAMAGER"
+					end
 				end
 			end