+++ Changelog.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 
+## [2.9.2] - 2024-10-23
+
+### Fixed
+
+- Added the new Blackrock Depths raid to the data.
+
+- [FilterPanel] There shouldn't be any more errors when Blizzard doesn't send all the class/spec data for the group listing.
+
+- [SearchPanel] There shouldn't be any more errors when Blizzard doesn't send all the class/spec data for the group listing.
+
+
+
 ## [2.9.1] - 2024-10-22
 
 ### Fixed